fix(): fix type for numeric into float

+++ src/city/model.rs
@@ -1,13 +1,14 @@
 use sqlx::FromRow;
 use serde::Serialize;
+use bigdecimal::BigDecimal;
 
 #[derive(Debug, FromRow, Serialize)]
 pub struct City {
-    pub id: i64,
+    pub id: i32,
     pub department_code: String,
     pub insee_code: String,
     pub zip_code: String,
     pub name: String,
-    pub lat: f64,
-    pub lon: f64,
+    pub lat: f32,
+    pub lon: f32,
 }
